requirements.txt
asgiref==3.9.1
autoflake==2.3.1
black==25.1.0
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
coverage==7.10.3
cryptography==45.0.6
distlib==0.4.0
django-anymail==13.0.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.1
filelock==3.18.0
flake8==7.3.0
identify==2.6.13
idna==3.10
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
mysqlclient==2.2.7
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.3.0
pycodestyle==2.14.0
pycparser==2.22
pyflakes==3.4.0
pytest-cov==6.2.1
pytest-django==4.11.1
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.33.1
PyGithub>=1.72
pytest
